psql check

+++ scripts/docker-entrypoint.sh
@@ -26,19 +26,20 @@ if [ "$1" = 'ofn' ]; then
     echo "===> Configuring Openfoodnetwork for production please wait..."
     sed -e "s#production:#${RAILS_ENV}:#" -e "s#.*adapter:.*#  adapter: postgresql#" -e "s#.*username:.*#  username: ${OFN_DB_USER}#" -e "s#.*password:.*#  password: ${OFN_DB_PASS}#" -e "s#.*database:.*#  database: ${OFN_DB}\n  host: ${OFN_DB_HOST}#" < ${OFN_DIR}/config/database.yml.pkgr > ${OFN_DIR}/config/database.yml
     cd ${OFN_DIR}
-    # populate database
-    echo "===> Running db:drop..."
-    bundle exec rake db:drop
-    echo "===> Running db:create..."
-    # psql -h postgresql -U $OFN_DB_USER -q -d $OFN_DB -c 'SELECT 1;' || bundle exec rake db:create
-    bundle exec rake db:create
-    echo "===> Running db:schema:load..."
-    bundle exec rake db:schema:load || echo "<== Schema already loaded..."
-    echo "===> Running db:migrate..."
-    bundle exec rake db:migrate || echo "<== already migrated..."
-    echo "===> Running db:seed..."
-    bundle exec rake db:seed || echo "<== Already seeded"
-
+    echo "==> Testing if database exists. if not, then populate database"
+    if ! psql -lqtA -h ${OFN_DB_HOST} -U ${OFN_DB_USER} | grep -qw ${OFN_DB} ; then
+      echo "===> Running db:drop..."
+      bundle exec rake db:drop
+      echo "===> Running db:create..."
+      # psql -h postgresql -U $OFN_DB_USER -q -d $OFN_DB -c 'SELECT 1;' || bundle exec rake db:create
+      bundle exec rake db:create
+      echo "===> Running db:schema:load..."
+      bundle exec rake db:schema:load || echo "<== Schema already loaded..."
+      echo "===> Running db:migrate..."
+      bundle exec rake db:migrate || echo "<== already migrated..."
+      echo "===> Running db:seed..."
+      bundle exec rake db:seed || echo "<== Already seeded"
+    fi
 
     # assets precompile
     echo "===> Running assets:precompile..."
